pyproject.toml
[tool.poetry]
name = "dynamic-dns-aws"
version = "0.0.0"
description = "Dynamic DNS that updates route53 configurations directly"
authors = ["Philip Couling <couling@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "dynamic_dns_aws.py", from = "app"}]

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1.16.43"
dnspython = "^2.0.0"
pyyaml = ">=5.1,<7"
click = "^8.1.3"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "pep440"
pattern = "^v?(?P<base>\\d+(\\.\\d+)*)"
bump = true

[build-system]
requires = ["poetry-core>=1.3.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry.scripts]
dynamic-dns-aws = "dynamic_dns_aws:main"